refactor: change structure of go-sgxvm rust code
